requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
audioop-lts==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colpali_engine==0.3.6
distro==1.9.0
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.17.0
fsspec==2024.12.0
GPUtil==1.4.0
gradio==5.13.1
gradio_client==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
openai==1.60.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
peft==0.14.0
pillow==11.1.0
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.3
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.7.0.dev20250125
torchaudio==2.6.0.dev20250125
torchvision==0.22.0.dev20250125
tqdm==4.67.1
transformers==4.46.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
